http-req: change package name
