+++ packages/http-req/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@lindeneg/core-request",
+  "name": "@lindeneg/http-req",
   "repository": "git@github.com:Lindeneg/lindeneg-npm-packages.git",
-  "homepage": "https://github.com/lindeneg/lindeneg-npm-packages/tree/master/packages/core-request",
+  "homepage": "https://github.com/lindeneg/lindeneg-npm-packages/tree/master/packages/http-req",
   "author": "lindeneg <christian@lindeneg.org>",
   "license": "MIT",
   "version": "1.0.0",
@@ -19,9 +19,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "gitHead": "4c3b992a653be2cdffeb05a50fe1c727853d39ce",
-  "devDependencies": {
-    "@lindeneg/types": "^1.1.2"
   }
 }
